requirements.txt
tensorflow
keras
pillow
numpy
Flask
Flask-Cors
numpy
Pillow
Flask
